+++ app/src/main/java/woowacourse/movie/domain/model/movie/Discountable.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.domain
+package woowacourse.movie.domain.model.movie
 
 interface Discountable {
     fun discount(money: Int): Int
