+++ app/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".view.fragments.MovieListFragment">
+    tools:context=".presentation.view.fragments.MovieListFragment">
 
     <ListView
         android:id="@+id/lv_movies"
