+++ app/src/main/res/layout/fragment_ticketing_result.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/white"
-    tools:context=".view.fragments.TicketingResultFragment">
+    tools:context=".presentation.view.fragments.TicketingResultFragment">
 
     <View
         android:id="@+id/ticketing_header_container"
