+++ app/src/test/java/woowacourse/movie/MovieDateTest.kt
@@ -2,7 +2,7 @@ package woowacourse.movie
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import woowacourse.movie.domain.MovieDate
+import woowacourse.movie.domain.model.movie.MovieDate
 import java.time.LocalDate
 
 class MovieDateTest {
